requirements.txt
crewai==0.80.0
crewai-tools==0.14.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
streamlit>=1.38.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-analytics-data>=0.18.0
google-ads>=24.0.0
sendgrid>=6.11.0
pandas>=2.1.0
plotly>=5.18.0
pyyaml>=6.0.0
schedule>=1.2.0
rich>=13.7.0
